renewable tin
this is just incase someone wants renewable tin! ever killed creeper will have a 50% chance of dropping raw tin. I will get the ore figured out, but for now if i were to go into survival with the goal of getting a griddle, it is now plausible! (I should also say i did this to make it easier to get allot of my custom items as most are electronic or electronic based to work and it would be a nightmare to explore simply for tin. i may adjust values as needed if it becomes overpowered, but i do hope someone who plays finds this usefull!)