tortoise egg
it takes 10 minutes to hatch. next up is making the tortoise actually lay the egg when bred with one another.

the turtles are made to test eggs. next up i plan on making more animals like birds and dinosaurs that hatch from eggs.

while the birds spawn naturauly; the dinosaurs wont and will be something that a player can choose to work for.

i will also try and make reasons to have dinosaurs and birds and also making the story last further then all that can be seen currently.

next up for sure though is making the tortoise lay the egg, then making tomatoes.