now the turtles will lay their eggs
when you breed 2 turtles, there will be a 30 second buffer before one of them will wander to a specific block to then lay an egg which will take a couple minutes to hatch into another tortoise.

note to self: dont place in a fenced area the turtles fly when the egg is placed.