a working solar panel
i also added a power meter so players can see how much power is in their machines, do know i didnt have to do this but i want to have something used for debugging and also if i wanted to have a screen i dont want it to be part of the screen. i will adjust it to also show a out of how much but that requires editation and i dont feel like doing that yet.