working item compressor
just adjustments to the scripts to allow for secondary input slots.

though i do still have it in the todo as i need to do the same to the grinder to make a use method for it so it can tae something like a glass bottle or bucket to make liquids or in certain cases allow for a method of carrying said things. i may try and make it so it doenst rely on those being in there so i can make something that isnt a liquid but it would do for now.