just need to make the hypsilophodon egg now
that will be simple but thought i'd add some small details and tweaks to make the hypsilophodon more final.

also next up after hypsilophodon will be to make a scarab mob set which will be attacking mobs in the prehistory and will only be found in specific biomes per scarab. i do plan on making a taming book for the dinos.