first entity base!
a entity base is a entity that has nothing to it but a basic outline; in minecraft (for which this first one extends!) something like tameable entity is something that contains all the basis for those made with it! in my case this is so custom scripts are broken down to what is needed!